requirements.txt

smbus
RPi.GPIO
pyglet
numpy>=1.18
pyaudio
scipy
VL53L1X
pytest-rerunfailures
pyserial
pysimplegui